+++ backend/packages/caching/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @cat-factory/caching
 
+## 0.14.2
+
+### Patch Changes
+
+- Updated dependencies [f1a6cb3]
+- Updated dependencies [cc17221]
+- Updated dependencies [889a497]
+- Updated dependencies [3605630]
+  - @cat-factory/kernel@0.226.0
+
 ## 0.14.1
 
 ### Patch Changes