+++ backend/packages/prompt-fragments/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @cat-factory/prompt-fragments
 
+## 0.15.49
+
+### Patch Changes
+
+- Updated dependencies [f1a6cb3]
+- Updated dependencies [cc17221]
+- Updated dependencies [889a497]
+- Updated dependencies [3605630]
+  - @cat-factory/contracts@0.224.0
+
 ## 0.15.48
 
 ### Patch Changes